feat：Extract `FeedCacheSpy` into a shared scope to remove duplication
