+++ EssentialFeedTests/Feed Cache/CacheFeedUseCaseTests.swift
@@ -132,44 +132,6 @@ final class CacheFeedUseCaseTests: XCTestCase {
         XCTAssertEqual(receivedError as NSError?, expectError, file: file, line: line)
     }
     
-    private class FeedStoreSpy: FeedStore {
-        enum ReceiveMessage: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        private(set) var receiveMessages = [ReceiveMessage]()
-        
-        private var deleteCompletions = [DeleteCompletion]()
-        private var insertionCompletions = [InsertionCompletion]()
-        
-        func deleteCachedFeed(completion: @escaping DeleteCompletion) {
-            deleteCompletions.append(completion)
-            receiveMessages.append(.deleteCachedFeed)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            deleteCompletions[index](error)
-        }
-        
-        func completeDeletionSuccessfully(at index: Int = 0) {
-            deleteCompletions[index](nil)
-        }
-        
-        func insert(_ items: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-            insertionCompletions.append(completion)
-            receiveMessages.append(.insert(items, timestamp))
-        }
-        
-        func completeInsertion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSuccessfully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-    }
-    
     private func uniqueImage() -> FeedImage {
         FeedImage(id: UUID(), description: "any", location: "any", imageURL: anyURL())
     }
