+++ EssentialFeedTests/Feed Cache/LoadFeedFromCacheUserCaseTests.swift
@@ -25,41 +25,4 @@ class LoadFeedFromCacheUserCaseTests: XCTestCase {
         return (sut: sut, store: store)
     }
     
-    private class FeedStoreSpy: FeedStore {
-        enum ReceiveMessage: Equatable {
-            case deleteCachedFeed
-            case insert([LocalFeedImage], Date)
-        }
-        
-        private(set) var receiveMessages = [ReceiveMessage]()
-        
-        private var deleteCompletions = [DeleteCompletion]()
-        private var insertionCompletions = [InsertionCompletion]()
-        
-        func deleteCachedFeed(completion: @escaping DeleteCompletion) {
-            deleteCompletions.append(completion)
-            receiveMessages.append(.deleteCachedFeed)
-        }
-        
-        func completeDeletion(with error: Error, at index: Int = 0) {
-            deleteCompletions[index](error)
-        }
-        
-        func completeDeletionSuccessfully(at index: Int = 0) {
-            deleteCompletions[index](nil)
-        }
-        
-        func insert(_ items: [LocalFeedImage], timestamp: Date, completion: @escaping InsertionCompletion) {
-            insertionCompletions.append(completion)
-            receiveMessages.append(.insert(items, timestamp))
-        }
-        
-        func completeInsertion(with error: Error, at index: Int = 0) {
-            insertionCompletions[index](error)
-        }
-        
-        func completeInsertionSuccessfully(at index: Int = 0) {
-            insertionCompletions[index](nil)
-        }
-    }
 }
